backend/feat/28/create_daos: [UNRELATED] rename screens to view
We renamed it to follow our new file structure.
