+++ app/src/main/java/io/github/mohamed/sallam/awb/view/devices/DevicesFragment.java
@@ -1,4 +1,4 @@
-package io.github.mohamed.sallam.awb.screens.repeatschedual;
+package io.github.mohamed.sallam.awb.view.devices;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import io.github.mohamed.sallam.awb.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link RepeatSchedualFragment#newInstance} factory method to
+ * Use the {@link DevicesFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class RepeatSchedualFragment extends Fragment {
+public class DevicesFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class RepeatSchedualFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public RepeatSchedualFragment() {
+    public DevicesFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class RepeatSchedualFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment RepeatSchedualFragment.
+     * @return A new instance of fragment DevicesFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static RepeatSchedualFragment newInstance(String param1, String param2) {
-        RepeatSchedualFragment fragment = new RepeatSchedualFragment();
+    public static DevicesFragment newInstance(String param1, String param2) {
+        DevicesFragment fragment = new DevicesFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class RepeatSchedualFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_repeat_schedual, container, false);
+        return inflater.inflate(R.layout.fragment_devices, container, false);
     }
 }
