+++ app/src/main/java/io/github/mohamed/sallam/awb/view/home/HomeFragment.java
@@ -1,4 +1,4 @@
-package io.github.mohamed.sallam.awb.screens.scheduals;
+package io.github.mohamed.sallam.awb.view.home;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import io.github.mohamed.sallam.awb.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SchedualsFragment#newInstance} factory method to
+ * Use the {@link HomeFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class SchedualsFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class SchedualsFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public SchedualsFragment() {
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class SchedualsFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment SchedualsFragment.
+     * @return A new instance of fragment HomeFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static SchedualsFragment newInstance(String param1, String param2) {
-        SchedualsFragment fragment = new SchedualsFragment();
+    public static HomeFragment newInstance(String param1, String param2) {
+        HomeFragment fragment = new HomeFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class SchedualsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_scheduals, container, false);
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 }
