+++ app/src/main/java/io/github/mohamed/sallam/awb/view/updategroup/UpdateGroupFragment.java
@@ -1,4 +1,4 @@
-package io.github.mohamed.sallam.awb.screens.updateschedual;
+package io.github.mohamed.sallam.awb.view.updategroup;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import io.github.mohamed.sallam.awb.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link UpdateSchedualFragment#newInstance} factory method to
+ * Use the {@link UpdateGroupFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class UpdateSchedualFragment extends Fragment {
+public class UpdateGroupFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class UpdateSchedualFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public UpdateSchedualFragment() {
+    public UpdateGroupFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class UpdateSchedualFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment UpdateSchedualFragment.
+     * @return A new instance of fragment UpdateGroupFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static UpdateSchedualFragment newInstance(String param1, String param2) {
-        UpdateSchedualFragment fragment = new UpdateSchedualFragment();
+    public static UpdateGroupFragment newInstance(String param1, String param2) {
+        UpdateGroupFragment fragment = new UpdateGroupFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class UpdateSchedualFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_update_schedual, container, false);
+        return inflater.inflate(R.layout.fragment_update_group, container, false);
     }
 }
