+++ app/src/main/java/io/github/mohamed/sallam/awb/view/updateschedual/UpdateSchedualFragment.java
@@ -1,4 +1,4 @@
-package io.github.mohamed.sallam.awb.screens.updategroup;
+package io.github.mohamed.sallam.awb.view.updateschedual;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import io.github.mohamed.sallam.awb.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link UpdateGroupFragment#newInstance} factory method to
+ * Use the {@link UpdateSchedualFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class UpdateGroupFragment extends Fragment {
+public class UpdateSchedualFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class UpdateGroupFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public UpdateGroupFragment() {
+    public UpdateSchedualFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class UpdateGroupFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment UpdateGroupFragment.
+     * @return A new instance of fragment UpdateSchedualFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static UpdateGroupFragment newInstance(String param1, String param2) {
-        UpdateGroupFragment fragment = new UpdateGroupFragment();
+    public static UpdateSchedualFragment newInstance(String param1, String param2) {
+        UpdateSchedualFragment fragment = new UpdateSchedualFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class UpdateGroupFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_update_group, container, false);
+        return inflater.inflate(R.layout.fragment_update_schedual, container, false);
     }
 }
