+++ app/src/main/res/layout/fragment_devices.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.devices.DevicesFragment">
+    tools:context=".view.devices.DevicesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
