+++ app/src/main/res/layout/fragment_home.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="HomeViewModel"
-            type="io.github.mohamed.sallam.awb.screens.home.HomeViewModel" />
+            type="io.github.mohamed.sallam.awb.view.home.HomeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".screens.home.HomeFragment">
+        tools:context=".view.home.HomeFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
