+++ app/src/main/res/layout/fragment_repeat_schedual.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.repeatschedual.RepeatSchedualFragment">
+    tools:context=".view.repeatschedual.RepeatSchedualFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
