+++ app/src/main/res/layout/fragment_scheduals.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.scheduals.SchedualsFragment">
+    tools:context=".view.scheduals.SchedualsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
