+++ app/src/main/res/layout/fragment_update_group.xml
@@ -5,13 +5,13 @@
     <data>
         <variable
             name="UpdateGroupViewModel"
-            type="io.github.mohamed.sallam.awb.screens.updategroup.UpdateGroupViewModel" />
+            type="io.github.mohamed.sallam.awb.view.updategroup.UpdateGroupViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".screens.updategroup.UpdateGroupFragment">
+        tools:context=".view.updategroup.UpdateGroupFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
