+++ app/src/main/res/layout/fragment_update_schedual.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.updateschedual.UpdateSchedualFragment">
+    tools:context=".view.updateschedual.UpdateSchedualFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
